requirements.txt
ortools
opencv-python
pytest
numpy
pyyaml
